Improving "import" implementation (#142)
* clean up examples

* update docs

* progress of implementing llvm.used

* added appendToUsed logic

* added used attribute for globals and functions

* added "used" attribute

* progress of implementing dll import code

* progress of adding logic to build obj file with __decls

* progress of adding declaration file to dll file

* update

* added code to use additional obj file with declarations for dll file

* added code for dll declaration

* update

* progress of adding dll decls for funcs without types

* progress

* added types for declarations in dll

* improved print types

* update

* fix for print declarations in dll

* added ignore for .d.ts in tests

* sync'd env vars for tests

* added output folder for .d.ts

* improving code to generare dll declarations

* progress of fixes for generating decls for dll

* clean up

* more fixes

* adding logic to get all symbols from dll

* added obj dumper for elf format

* update comments

* improved code

* fixes to loading symbols

* finishing code to load decl from dlls

* final bit

* fixes to string allocs

* fixes

* fixing issue with binding to varaibles from dll

* progress of implementing dll import for vars

* progress of implementing Dll vars

* fix for linux compile

* added type print

* refactoring globalOp to support Linkage and debug info

* added namedloc for GlobalOp

* fixes

* refactoring

* added globalop debug info

* fixes for local linkage of exporting globals

* fix for wrong linkage for export import vars

* final fix for debug info

* update

* update build files

+++ docs/how/cmake_winapp/src/appwindow.ts
@@ -8,7 +8,7 @@ export class AppWindow {
         this.handler_window = create_window('Hello World!', parent_handler_window, this.onMessage);
     }
 
-    protected onMessage(uMsg: uint32_t, wParam: uint64_t, lParam: uint64_t): intptr_t {
+    protected onMessage(uMsg: uint32_t, wParam: uint64_t, lParam: uint64_t): uint32_t {
         switch (uMsg) {
             case Messages.Close:
                 close_window(0);
